flow_manager/requirements.txt
ccxt==4.4.24
python-telegram-bot==13.15
pydantic==2.8.2
